### 2025-06-24 #290 @editorialbot
[PRE REVIEW]: XAI-MICROPYTHON: Explainable Artificial Intelligence with MicroPython
**Submitting author:** @statistical-thinking (<a href="http://orcid.org/0000-0003-1011-5517">Prof. Dr. habil. Dennis Klinkhammer</a>)
**Repository:** https://github.com/statistical-thinking/xai-micropython
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
