### 2026-04-16 #290 @statistical-thinking
@labarba Dear EiC, unfortunately, we haven’t been able to find any reviewers for my open-source framework over the past ten months. I’d like to emphasize once again that it essentially covers only the mathematical and statistical foundations of machine learning and deep learning, which is why many scientists should be well-suited to review it. Since I’ve released an updated version of the tool, I’d appreciate an update. Maybe we could bring in some reviewers from JOSS?
