### 2026-05-20 #290 @labarba
Hi @statistical-thinking 👋  – thank you for your continued patience.
Right now we have 72 in-progress submissions, of which only 13 have progressed to review. I estimate we need more than 100 volunteer reviewers to move forward with the current backlog. If you want to help, please look over the [submissions in pre-review,](https://github.com/openjournals/jose-reviews/issues) and volunteer for one in your area of interest!
