### 2026-06-15 #379 @editorialbot
[REVIEW]: XAI-MICROPYTHON: Explainable Artificial Intelligence with MicroPython
**Submitting author:** @statistical-thinking (<a href="http://orcid.org/0000-0003-1011-5517">Prof. Dr. habil. Dennis Klinkhammer</a>)
**Repository:** https://github.com/statistical-thinking/xai-micropython
**Branch with paper.md** (empty if default branch):
**Version:** v1.5.0
**Paper kind:** software
