### 2026-06-24 bfecf01
Update paper.md
