### 2026-07-02 54dbfb2
Update xai_micropython.py
