### 2026-07-02 2e57730
Update xai_micropython.py
