### 2026-07-02 e6669be
Update xai_micropython.py
