### 2026-07-02 eb811eb
Update xai_micropython.py
