### 2026-07-02 83457e7
Update xai_micropython.py
