### 2026-07-02 9eb4826
Update xai_micropython.py
