+++ README.md
@@ -9,6 +9,9 @@ Prof. Dr. habil. Dennis Klinkhammer
 # Requirements
 MicroPython capable microcontrollers (or devices) or browser (e.g. Firefox).
 
+# Educational Motivation
+XAI-MicroPython is an **educational framework** for developing and executing explainable machine learning and deep learning algorithms on MicroPython-compatible microcontrollers. The framework is designed to **support teaching and learning of the fundamentals of artificial intelligence** on resource-constrained hardware. It introduces **regression, classification, clustering, and factor analysis** from machine learning, as well as the **architecture and operation of artificial neural networks** from deep learning. By running directly on microcontrollers, learners can engage with the underlying mathematical and statistical concepts **without depending on high-level machine learning frameworks**.
+
 # Live-Demo in Browser
 XAI-MICROPYTHON can be used directly in the browser as **Web Version** or **Jupyer Notebook**! The easiest way is the web version, which explains all the code elements from statistics to machine learning and deep learning without any prerequisites: [WEB VERSION](https://statistical-thinking.de/xai/index.html). And the following link opens **JupyterLite** and the **Jupyter Notebook** with the name **xai-micropython.ipynb** can be called up by double-clicking. There is a concise explanation for each cell with MicroPython code, which can be extended as required in your own teaching settings. To execute the code, simply mark the corresponding cell and click on **Run Selected Cell**. Alternatively, all cells can be executed by clicking on **Run All Cells**. This is the link to the live-demo: [JUPYTER NOTEBOOK](https://statistical-thinking.github.io/xai-micropython)
 
