### 2026-07-02 29699a4
Update xai_micropython.py
