### 2026-07-02 f8e03c9
Update xai_micropython.py
