### 2026-08-21 #379 @DiegoAscanio
Please accept my apologies for the delay in completing this review.

After a careful reading, I did not identify any issues that would make this work unsuitable for publication in JOSE. Therefore, I recommend its acceptance.
