### 2026-08-21 #379 @statistical-thinking
Dear @DiegoAscanio, that's awesome news, thank you!

As I understand it, the only thing **missing is the checkmark next to "Performance Claims"**, since my XAI Framework doesn't make any performance claims :-)

Otherwise, JOSE won't be able to finish the review - is that right, @stats-tgeorge?
