### 2026-08-24 #379 @stats-tgeorge
@statistical-thinking, please see the short checklist above.
