### 2026-08-24 #379 @statistical-thinking
Dear @stats-tgeorge, I've checked the Additional Author Tasks as follows:

-> authors and affiliations are correct and this is my **ORCIDs** (https://orcid.org/0000-0003-1011-5517)
-> the **latest release** of XAI-MICROPYTHON is **1.5.0** (https://github.com/statistical-thinking/xai-micropython)
-> released on **figshare** (https://doi.org/10.6084/m9.figshare.33324921)
-> title and author list in the archive **do match** those in the JOSE paper
->  **license** listed for the archive is the same as the software license
