### 2026-08-24 #379 @stats-tgeorge
@statistical-thinking  fast work! Did you see the issue I opened on GitHub?
