+++ README.md
@@ -30,6 +30,17 @@ XAI-MICROPYTHON can be used directly in the browser as **Web Version** or **Jupy
 # Execution on a Microcontroller
 XAI-MICROPYTHON is so lightweight that it can even run on a microcontroller such as the **Raspberry Pi Pico 2**. If this is your first project with this kind of hardware, I recommend checking out the official [PICO DOCUMENTATION](https://pip-assets.raspberrypi.com/categories/610-raspberry-pi-pico/documents/RP-008276-DS-1-getting-started-with-pico.pdf). And if this is your first time with MicroPython (or Python as general purpose programming language), there is an official [PYTHON DOCUMENTATION](https://pip-assets.raspberrypi.com/categories/610-raspberry-pi-pico/documents/RP-008355-DS-1-raspberry-pi-pico-python-sdk.pdf) as well. In order to get started, simply transfer **xai_micropython.py** to the microcontroller via [THONNY](https://thonny.org/), for example, and execute it. In addition, devices such as the [PICOCALC](https://www.clockworkpi.com/picocalc) can be used to **create your own learning devices** with XAI-MICROPYTHON. 
 
+# Adoption and Use in Teaching
+XAI-MICROPYTHON is designed for **flexible adoption in educational settings**. It can be used without dedicated hardware through the browser-based web version or Jupyter Notebook, while the same concepts and algorithms can also be explored directly on MicroPython-compatible microcontrollers.
+
+Educators can adopt the materials as a **complete learning sequence from statistical foundations to machine learning and deep learning**, or select individual examples and algorithms according to the objectives of their course. The concise explanations and executable code examples can be used for **instructor-led demonstrations**, **guided exercises**, or **independent experimentation** by learners. Code examples can also be **modified and extended to support further assignments** and project-based learning.
+
+In higher education teaching, XAI-MICROPYTHON has been used for assignments in which students extend and modify the provided implementations. In the machine learning section, **for example, students can implement additional clustering algorithms** by translating their mathematical formulation into MicroPython code. In the deep learning section, **students can manually modify the architecture of an artificial neural network or implement additional activation functions**. These activities encourage learners to **connect mathematical concepts directly with their computational implementation** and to explore how changes at the algorithmic level affect the resulting models.
+
+A **low-threshold introduction** can begin with the web version, which provides explanations and executable examples without requiring a local installation. The Jupyter Notebook offers **additional opportunities for interactive experimentation with individual code cells**. For hardware-oriented teaching settings, the MicroPython implementation can subsequently be **transferred to a compatible microcontroller**, allowing learners to apply the same computational concepts on resource-constrained hardware.
+
+This flexibility enables educators to adapt XAI-MICROPYTHON to **different levels of prior programming experience**, available technical infrastructure, and teaching formats.
+
 # Community Guidelines
 XAI-MICROPYTHON is designed to be used in (academic) teaching without any technical requirements. A preselection of statistical functions as well as machine learning and deep learning methods are available for this purpose. However, their functionality and the addition of further algorithms and methods are a primary goal of XAI-MICROPYTHON. Interested users can therefore contact me at any time in order to...
 
