### 2026-08-24 d74fae6
Update paper.md
