### 2026-08-24 d4c341b
Update paper.md
