+++ paper/paper.md
@@ -29,6 +29,6 @@ XAI-MICROPYTHON draws on well-known exercise data sets and common methods to giv
 The methods presented with XAI-MICROPYTHON can actually be easily implemented in Python using libraries such as SciKit-Learn. However, the aim of XAI-MICROPYTHON is not only to call up and apply already existing functions, but also to understand their methodological foundations via programming from scratch in MicroPython and without recourse to previous libraries [@Kong2022-sk]. A previous investigation of skill requirements in artificial intelligence and machine learning job advertisements [@Verma2022-ie] served as a guide for the development of XAI-MICROPYTHON. It is intended to provide insights into basic statistical principles, some machine learning algorithms as well as the structure and functioning of neural networks as basis for artificial intelligence [@Schmidt2020-ak; @Frank2020-rw; @Scherer2016-fy]. Accordingly, the learning objectives are to enable learners to understand the mathematical and statistical foundations of artificial intelligence, to translate these foundations into transparent MicroPython implementations, and to independently modify and extend the implemented algorithms and models. Since XAI-MICROPYTHON is based on the MicroPython programming language, it can be used on microcontrollers like the Raspberry Pi Pico 2 [@Sakr2021-mc] or it can be called up directly via JupyterLite as a live demo in the browser. Especially the step-by-step visibility in MicroPython helps users to identify and understand the mathematical and statistical foundations of artificial intelligence [@Kong2022-sk]. This is in accordance with the primary goals of explainable artificial intelligence [@Haque2023-du; @Schmidt2020-ak] and provides insights into artificial intelligence related responsibilities [@Collier2024-ld; @Frank2020-rw] as well as practical experiences [@Li2021-bn]. Therefore, XAI-MICROPYTHON shall promote necessary innovation and problem-solving skills in accordance with the scientific discourse [@Verma2022-ie].
 
 # Basic elements
-XAI-MICROPYTON introduces to mean, variance, standard deviation, covariance and correlation as elements of univariate and bivariate statistics, which are necessary for the implementation of machine learning algorithms like single and multiple linear regressions. Simple classification tasks will be highlighted via multiple logistic regression, followed by k-Means as algorithm, as well as the concept of the within cluster sum of squares, for the demonstration of clustering. Furthermore, factor analysis will be implemented in MicroPython in order to demonstrate dimensionality reduction. In particular, the division into four different machine learning application areas is based on the classification of SciKit-Learn (regression, classification, clustering and dimensionality reduction). Finally, a pre-trained neural network and a self-learning neural network will be implemented in MicroPython in order to highlight some common activation functions as well as the underlying principles of layers and neurons. Therefore, XAI-MICROPYTHON is an underlying framework, which can be extended independently by (academic) teachers with further algorithms and methods and is intended to grow via the community guidelines provided.
+XAI-MICROPYTHON introduces to mean, variance, standard deviation, covariance and correlation as elements of univariate and bivariate statistics, which are necessary for the implementation of machine learning algorithms like single and multiple linear regressions. Simple classification tasks will be highlighted via multiple logistic regression, followed by k-Means as algorithm, as well as the concept of the within cluster sum of squares, for the demonstration of clustering. Furthermore, factor analysis will be implemented in MicroPython in order to demonstrate dimensionality reduction. In particular, the division into four different machine learning application areas is based on the classification of SciKit-Learn (regression, classification, clustering and dimensionality reduction). Finally, a pre-trained neural network and a self-learning neural network will be implemented in MicroPython in order to highlight some common activation functions as well as the underlying principles of layers and neurons. Therefore, XAI-MICROPYTHON is an underlying framework, which can be extended independently by (academic) teachers with further algorithms and methods and is intended to grow via the community guidelines provided.
 
 # References
