### 2026-08-24 #379 @statistical-thinking
> [@statistical-thinking](https://github.com/statistical-thinking) fast work! Did you see the issue I opened on GitHub?

@stats-tgeorge, I have **added very detailed information to the README** file to better describe these two points and have **briefly addressed them in the paper** accordingly...
