+++ README.md
@@ -1,7 +1,7 @@
 ![title](images/xai_micropython.png)
 
 # XAI-MICROPYTHON
-Explainable Artificial Intelligence (XAI) with MicroPython - a comprehensice codebook that covers relevant topics in AI **from statistical basics to machine learning and deep learning** and makes them **executable in MicroPython** (suitable for microcontrollers and devices like the PicoCalc) and live in your Browswer.
+Explainable Artificial Intelligence (XAI) with MicroPython - a comprehensive codebook that covers relevant topics in AI **from statistical basics to machine learning and deep learning** and makes them **executable in MicroPython** (suitable for microcontrollers and devices like the PicoCalc) and live in your browswer.
 
 # Author
 Prof. Dr. habil. Dennis Klinkhammer
