+++ README.md
@@ -10,7 +10,7 @@ Prof. Dr. habil. Dennis Klinkhammer
 MicroPython capable microcontrollers (or devices) or browser (e.g. Firefox).
 
 # Educational Motivation
-XAI-MICROPYTHON is an **educational framework** for developing and exploring explainable machine learning and deep learning algorithms on MicroPython-compatible microcontrollers. The framework is designed to **support teaching and learning of the fundamentals of artificial intelligence** on resource-constrained hardware. It introduces **regression, classification, clustering, and factor analysis** from machine learning, as well as the **architecture and operation of artificial neural networks** from deep learning. By running directly on microcontrollers, learners can engage with the underlying mathematical and statistical concepts **without depending on high-level machine learning frameworks**.
+XAI-MICROPYTHON is an **educational framework** for developing and exploring explainable machine learning and deep learning algorithms on MicroPython-compatible microcontrollers. The framework is designed to **support the teaching and learning of the fundamentals of artificial intelligence** on resource-constrained hardware. It introduces **regression, classification, clustering, and factor analysis** as machine learning methods, as well as the **architecture and operation of artificial neural networks** in deep learning. By running directly on microcontrollers, learners can engage with the underlying mathematical and statistical concepts **without relying on high-level machine learning frameworks**.
 
 # Learning Objectives
 XAI-MICROPYTHON aims to make the **mathematical and computational foundations of artificial intelligence accessible and transparent**. After working with the learning materials, learners should be able to:
