+++ README.md
@@ -25,10 +25,10 @@ XAI-MICROPYTHON aims to make the **mathematical and computational foundations of
 The individual examples are designed to progress from statistical foundations to machine learning and deep learning while keeping the implementation sufficiently transparent for learners to **inspect, modify, and experiment with the underlying code**.
 
 # Live-Demo in Browser
-XAI-MICROPYTHON can be used directly in the browser as either a **web version** or a **Jupyter Notebook**. The easiest way to get started is the web version, which explains all code elements, from statistics to machine learning and deep learning, without requiring any prior knowledge: [WEB VERSION](https://statistical-thinking.de/xai/index.html). The following link opens **JupyterLite**, where the Jupyter Notebook named `xai-micropython.ipynb` can be opened by double-clicking it. Each cell containing MicroPython code includes a **concise explanation** and can be extended as required for your own teaching purposes. To execute the code, simply select the corresponding cell and click on `Run Selected Cell`. Alternatively, all cells can be executed by clicking on `Run All Cells`. Here is the link to the live demo: [JUPYTER NOTEBOOK](https://statistical-thinking.github.io/xai-micropython).
+XAI-MICROPYTHON can be used directly in the browser as either a **web version** or a **Jupyter Notebook**. The easiest way to get started is the web version, which explains all code elements, from statistics to machine learning and deep learning, without requiring any prior knowledge: [WEB VERSION](https://statistical-thinking.de/xai/index.html). The following link opens **JupyterLite**, where the Jupyter Notebook named `xai-micropython.ipynb` can be opened by double-clicking it. Each cell containing MicroPython code includes a **concise explanation** and **can be extended as required for your own teaching purposes**. To execute the code, simply select the corresponding cell and click on `Run Selected Cell`. Alternatively, all cells can be executed by clicking on `Run All Cells`. Here is the link to the live demo: [JUPYTER NOTEBOOK](https://statistical-thinking.github.io/xai-micropython).
 
 # Execution on a Microcontroller
-XAI-MICROPYTHON is so lightweight that it can even run on a microcontroller such as the **Raspberry Pi Pico 2**. If this is your first project with this kind of hardware, I recommend checking out the official [PICO DOCUMENTATION](https://pip-assets.raspberrypi.com/categories/610-raspberry-pi-pico/documents/RP-008276-DS-1-getting-started-with-pico.pdf). And if this is your first time with MicroPython (or Python as general purpose programming language), there is an official [PYTHON DOCUMENTATION](https://pip-assets.raspberrypi.com/categories/610-raspberry-pi-pico/documents/RP-008355-DS-1-raspberry-pi-pico-python-sdk.pdf) as well. In order to get started, simply transfer **xai_micropython.py** to the microcontroller via [THONNY](https://thonny.org/), for example, and execute it. In addition, devices such as the [PICOCALC](https://www.clockworkpi.com/picocalc) can be used to **create your own learning devices** with XAI-MICROPYTHON. 
[…]
