### 2026-08-25 bfb4cbd
Update paper.md
