+++ paper/paper.md
@@ -22,13 +22,13 @@ bibliography: paper.bib
 XAI-MICROPYTHON is a MicroPython based framework which emphasizes the fundamentals of statistics, machine learning and deep learning, since they are increasingly driving innovation across various fields [@LeCun2015-hn]. These fundamentals can also be transferred on microcontrollers [@Ray2022-oq], for educational purpose or e.g. in order to process sensor data [@Cioffi2020-lq]. Since microcontrollers have limited computational resources, a particularly efficient implementation of these methods is required [@Delnevo2023-ux; @Wulfert2024-qf]. However, it is the efficient implementation that provides insights into the fundamentals of statistics, machine learning and deep learning in terms of explainable artificial intelligence that makes this approach also relevant for educational purpose [@Haque2023-du; @Meske2022-ke]. Coding from scratch in MicroPython allows for a streamlined and tailored approach [@Delnevo2023-ux] and provides users with transparent insights into underlying principles and the resulting code, which makes XAI-MICROPYTHON suitable for didactic application [@Collier2024-ld; @Verma2022-ie; @Scherer2016-fy].
 
 # Target group
-XAI-MICROPYTHON draws on well-known exercise data sets and common methods to give (academic) teachers a quick start in designing their own teaching materials. In addition, XAI-MICROPYTHON can of course also be used by self-learners. In higher education teaching, XAI-MICROPYTHON has been used for assignments in which students extend the provided implementations, for example by translating additional clustering algorithms from their mathematical formulation into MicroPython code, modifying the architecture of neural networks, or implementing additional activation functions. Educators can adopt the framework as a complete learning sequence or select individual examples according to their teaching objectives, using either the browser-based environment or MicroPython-compatible microcontrollers.
+XAI-MICROPYTHON draws on well-known exercise data sets and common methods to give (academic) teachers a quick start in designing their own teaching materials [@Collier2024-ld; @Haque2023-du]. In addition, XAI-MICROPYTHON can of course also be used by self-learners. In higher education teaching, XAI-MICROPYTHON has been used for assignments in which students extend the provided implementations, for example by translating additional clustering algorithms from their mathematical formulation into MicroPython code, modifying the architecture of neural networks, or implementing additional activation functions. Educators can adopt the framework as a complete learning sequence or select individual examples according to their teaching objectives [@Verma2022-ie], using either the browser-based environment or MicroPython-compatible microcontrollers.
 
 # Statement of need
 
[…]
