### 2026-08-25 #379 @statistical-thinking
Just a **small update on the references** since my previous update regarding the learning objectives as well as the adoption and use in teaching. I also found and **corrected the last typo** in the README :-)
