### 2026-08-25 #379 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
