### 2026-08-25 #379 @statistical-thinking
Thank you very much, @stats-tgeorge, as well as @omendezmorales and @DiegoAscanio!

Dear @labarba, if you're around, could you send me the link for new reviewers again? I'd really like to sign up!
